Move non-script files to common/ subdir
