+++ mozapkpublisher/common/exceptions.py
