+++ mozapkpublisher/common/main_logging.py
