+++ mozapkpublisher/common/utils.py
